First pass at fixing Revert To Status Branch
The issue is that paths are in different formats and comparisons don't work (see comments). This works for the base game but requires more testing for plugin content. Matt has also recommended to look into "CreateStandardFilename" and "MakeStandardFilename" as a better alternative, but I want to get this "working" version in first.