Disable changelists and the additional "Submit Content" button in UEGitPlugin
The the UE git implementation doesn't support changelists because that isn't a feature that git supports. Recently a feature was added to UEGitPlugin that fakes changelists so changes, specifically One File Per Actor changes, can be staged and committed via an external git client. As we have auto stage enabled it renders this functionality broken.

Because this is default enabled behaviour in the UEGitPlugin they had to add a custom "Submit Content" button as UE hides it if changelists are enabled. By disabling changelists we hide the "View Changes" menu item but gain the built in "Submit Content" button. We need to disable the "Submit Content" button added in the UEGitPlugin so we don't double up.